.history/requirements_20250626155640.txt
# Core dependencies
yfinance>=0.2.63
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# LLM and AI
transformers>=4.46.0
torch>=2.4.0
ollama>=0.1.7
langchain>=0.2.0
langchain-community>=0.2.0

# Web scraping and data collection
firecrawl>=0.1.0
selenium>=4.15.0
webdriver-manager>=4.0.0
scrapy>=2.11.0

# Financial analysis
ta-lib>=0.4.28
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Data processing and storage
sqlite3
json5>=0.9.0
pyyaml>=6.0.0

# API and networking
aiohttp>=3.9.0
websockets>=12.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0
schedule>=1.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
